+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file. The format 
 - Grains embedded into other grains are now identified in the splinegon creation algorithm. [4cef7ec](https://github.com/CsatiZoltan/CristalX/commit/4cef7ecbbe4c3fc5cddb9947b29b3d0b454eecb8), [6c4ad0a](https://github.com/CsatiZoltan/CristalX/commit/6c4ad0aff2c4dae49fb648656a94b14436925bed)
 - Computation of the infinitesimal and Green-Lagrange strain tensors for DIC displacement field. [7eaa0b5](https://github.com/CsatiZoltan/CristalX/commit/7eaa0b523c0bc7e7a585f955aa8c31206cff751a), [9e5466e](https://github.com/CsatiZoltan/CristalX/commit/9e5466ebe1f4ab19defb9e91ff2ed8a961b7044b)
 - Computation of the equivalent von Mises strain. [2b9b0be](https://github.com/CsatiZoltan/CristalX/commit/2b9b0be36cf38f5d1683821573646a65ca541160)
+- The README file shows how to cite our paper. [#30](https://github.com/CsatiZoltan/CristalX/issues/30)
+- Added a new segmented microstructure ([6b539ba](https://github.com/CsatiZoltan/CristalX/commit/6b539ba356bb2e27f6dca4fbc299ca90635d3ab2)) and the corresponding DIC measurements ([3c845e7](https://github.com/CsatiZoltan/CristalX/commit/3c845e7e180a75acb2ea27404d75c308d35ca786)).
 
 
 ### Deprecated